Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
-
-float main() {
+int main() {
     int a,b;
     char opt;
     scanf("%d %d",&a,&b);
